Another attempt at windows fix. #107

+++ src/RcppExports.cpp
@@ -17,13 +17,13 @@ BEGIN_RCPP
 END_RCPP
 }
 // read_connection_
-RawVector read_connection_(RObject con, size_t chunk_size);
+RawVector read_connection_(RObject con, int chunk_size);
 RcppExport SEXP readr_read_connection_(SEXP conSEXP, SEXP chunk_sizeSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
     Rcpp::traits::input_parameter< RObject >::type con(conSEXP);
-    Rcpp::traits::input_parameter< size_t >::type chunk_size(chunk_sizeSEXP);
+    Rcpp::traits::input_parameter< int >::type chunk_size(chunk_sizeSEXP);
     __result = Rcpp::wrap(read_connection_(con, chunk_size));
     return __result;
 END_RCPP
